+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## unreleased
+* Deprecate `masterpass_card` and `amex_checkout_card` payment methods
+
 ## 4.3.0
 * Add validation error code `Transaction.ProductSkuIsInvalid`
 * Add 'RiskThreshold' gateway rejection reason