+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## unreleased
+## 4.4.0
 * Deprecate `masterpass_card` and `amex_checkout_card` payment methods
 * Fix issue where `transaction.credit` could not be called using a gateway object